getout settigs from config.h, because its now a own class
