+++ .gitignore
@@ -3,5 +3,5 @@
 .vscode/c_cpp_properties.json
 .vscode/launch.json
 .vscode/ipch
-*config.h
 /.pio/build
+/src/config/settings.h
